refactor: changing architecture to domain driven
