+++ package.json
@@ -24,6 +24,7 @@
         "@emotion/styled": "^11.10.6",
         "@mui/icons-material": "^5.11.11",
         "@mui/material": "^5.11.12",
+        "@tanstack/react-query": "^4.28.0",
         "date-fns": "^2.29.3",
         "i18next": "^22.4.11",
         "react": "^18.2.0",
